backend/requirements.txt
# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.35
asyncpg>=0.30.0
alembic>=1.14.0
psycopg2-binary>=2.9.10

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.28.0

# Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# Google AI
google-generativeai>=0.8.0
google-genai>=1.0.0

# Utils
python-dotenv>=1.0.1
tenacity>=9.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
